// src/common/enums/bff-error-codes.enum.ts
export enum BFFErrorCodes {
    USER_NOT_FOUND = 'USER_NOT_FOUND',
    USER_ALREADY_REGISTERED = 'USER_ALREADY_REGISTERED',
    USER_ERROR_REGISTER = 'USER_ERROR_REGISTER',
    INVALID_CODE_PROVIDED = 'INVALID_CODE_PROVIDED',
    EXPIRATE_CODE_PROVIDED = 'EXPIRATE_CODE_PROVIDED',
    USER_NOT_AUTHORIZE = 'USER_NOT_AUTHORIZE',
    ERROR_LOGIN_USER = 'ERROR_LOGIN_USER',
    ERROR_SIGN_OUT = 'ERROR_SIGN_OUT',
    ERROR_CHANGE_PASSWORD_CODE = 'ERROR_CHANGE_PASSWORD_CODE',
    ERROR_GROUP_PROJECT_EXIST = 'ERROR_GROUP_PROJECT_EXIST',
    ERROR_READ_GROUP_PROJECT = 'ERROR_READ_GROUP_PROJECT',
    ERROR_ALL_GROUP_PROJECT = 'ERROR_ALL_GROUP_PROJECT',
    ERROR_UPDATE_GROUP_PROJECT = 'ERROR_UPDATE_GROUP_PROJECT',
    ERROR_DELETE_GROUP_PROJECT = 'ERROR_DELETE_GROUP_PROJECT',
    ERROR_DELETE_GROUP_PROJECT_ACTIVE = 'ERROR_DELETE_GROUP_PROJECT_ACTIVE',
    ERROR_PROFILE_EXIST = 'ERROR_PROFILE_EXIST',
    ERROR_GET_PROFILES = 'ERROR_GET_PROFILES'
  };