// src/common/enums/bff-success-general.enum.ts
export enum BFFSuccessCodes {
  SUCCESS_SIGN_OUT = "SUCCESS_SIGN_OUT",
  SUCCESS_GROUP_PROJECTS = "SUCCESS_GROUP_PROJECTS",
  SUCCESS_UPDATE_GROUP_PROJECTS = "SUCCESS_UPDATE_GROUP_PROJECTS",
  SUCCESS_DELETE_GROUP_PROJECTS = "SUCCESS_DELETE_GROUP_PROJECTS",
  SUCCESS_ACTIVE_GROUP_PROJECTS = "SUCCESS_ACTIVE_GROUP_PROJECTS",
  SUCCESS_PROFILES = "SUCCESS_PROFILES"
};
export enum FeedbackSuccessCode {
  SUCCESS_FEEDBACK_CODE = "SUCCESS_FEEDBACK_CODE",
  UPDATE_FEEDBACK_CODE = "UPDATE_FEEDBACK_CODE"
};
export enum GenerateToken {
  SUCCESS_GENERATE_TOKEN = "SUCCESS_GENERATE_TOKEN"
};